+++ raiden-dapp/src/components/account/Withdrawal.vue
@@ -121,7 +121,6 @@
 <script lang="ts">
 import { Component, Mixins } from 'vue-property-decorator';
 import { mapState, mapGetters } from 'vuex';
-import { Tokens } from '@/types';
 import { Token } from '@/model/types';
 import AddressDisplay from '@/components/AddressDisplay.vue';
 import BlockieMixin from '@/mixins/blockie-mixin';
@@ -129,8 +128,8 @@ import AmountDisplay from '@/components/AmountDisplay.vue';
 import ActionButton from '@/components/ActionButton.vue';
 import RaidenDialog from '@/components/dialogs/RaidenDialog.vue';
 import Spinner from '@/components/icons/Spinner.vue';
-import { Zero } from 'ethers/constants';
 import uniqBy from 'lodash/uniqBy';
+import { Zero } from 'ethers/constants';
 import { BigNumber } from 'ethers/utils';
 
 @Component({
@@ -142,12 +141,12 @@ import { BigNumber } from 'ethers/utils';
     Spinner,
   },
   computed: {
-    ...mapState(['tokens', 'raidenAccountBalance']),
-    ...mapGetters(['udcToken']),
+    ...mapState(['raidenAccountBalance']),
+    ...mapGetters(['udcToken', 'allTokens']),
   },
 })
 export default class Withdrawal extends Mixins(BlockieMixin) {
-  tokens!: Tokens;
+  allTokens!: Token[];
   raidenAccountBalance!: string;
   udcToken!: Token;
   balances: Token[] = [];
@@ -156,15 +155,23 @@ export default class Withdrawal extends Mixins(BlockieMixin) {
   withdraw: Token | null = null;
 
   async mounted() {
-    this.balances = await this.$raiden.getRaidenAccountBalances();
+    const allTokens = uniqBy(
+      this.allTokens.concat(this.udcToken),
+      (token) => token.address
+    );
+    const updatedTokenBalances = await Promise.all(
+      allTokens.map(async (token) => ({
+        ...token,
+        balance: await this.$raiden.getTokenBalance(
+          token.address,
+          this.$store.state.defaultAccount
+        ),
+      }))
+    );
 
-    if ((this.udcToken.balance as BigNumber).gt(Zero)) {
-      this.balances = uniqBy(
-        [...this.balances].concat(this.udcToken),
-        (token) => token.address
-      );
-    }
-
+    this.balances = updatedTokenBalances.filter((token) =>
+      (token.balance as BigNumber).gt(Zero)
+    );
     this.loading = false;
   }