+++ raiden-dapp/tests/unit/components/account/withdrawal.spec.ts
@@ -1,146 +1,101 @@
-import RaidenDialog from '@/components/dialogs/RaidenDialog.vue';
-
-jest.mock('@/services/raiden-service');
 jest.mock('@/i18n', () => jest.fn());
-
+import flushPromises from 'flush-promises';
 import { mount, Wrapper } from '@vue/test-utils';
 import Vue from 'vue';
 import Vuex from 'vuex';
 import Vuetify from 'vuetify';
-import store from '@/store';
 import { $identicon } from '../../utils/mocks';
+import { BigNumber } from 'ethers/utils';
 import Withdrawal from '@/components/account/Withdrawal.vue';
-import RaidenService from '@/services/raiden-service';
-import Mocked = jest.Mocked;
-import { parseUnits } from 'ethers/utils';
-import { Zero } from 'ethers/constants';
+import RaidenDialog from '@/components/dialogs/RaidenDialog.vue';
 
-Vue.use(Vuetify);
 Vue.use(Vuex);
+Vue.use(Vuetify);
+
+let vuetify: typeof Vuetify;
+let $raiden: any;
+
+const tokenBalance = new BigNumber('1000000');
+const tokenNoBalance = new BigNumber('0');
+const raidenBalance = '1000000';
+const raidenNoBalance = '0';
+
+function createWrapper(
+  tokenBalance: BigNumber,
+  raidenAccountBalance: string
+): Wrapper<Withdrawal> {
+  $raiden = {
+    getTokenBalance: jest.fn().mockResolvedValue(tokenBalance),
+    transferOnChainTokens: jest.fn(),
+  };
+
+  const state = {
+    raidenAccountBalance,
+  };
+
+  const getters = {
+    udcToken: () => {
+      return {
+        address: '0xuserdeposittoken',
+        balance: new BigNumber('0'),
+      };
+    },
+    allTokens: () => [],
+  };
+
+  const store = new Vuex.Store({ state, getters });
+
+  return mount(Withdrawal, {
+    vuetify,
+    store,
+    stubs: ['v-dialog', 'i18n'],
+    mocks: {
+      $identicon: $identicon(),
+      $t: (msg: string) => msg,
+      $raiden,
+    },
+  });
+}
+
+beforeEach(() => {});
 
 describe('Withdrawal.vue', () => {
-  let wrapper: Wrapper<Withdrawal>;
-  let vuetify: typeof Vuetify;
-  let $raiden: Mocked<RaidenService>;
+  test('can withdraw tokens', async () => {
+    const wrapper = createWrapper(tokenBalance, raidenBalance);
+    await flushPromises();
 
-  function createWrapper(raidenAccountBalance: string = '') {
-    const state = {
-      tokens: {},
-      raidenAccountBalance,
-    };
+    wrapper.find('.withdrawal__tokens__button').trigger('click');
+    await wrapper.vm.$nextTick();
 
-    const getters = {
-      balance: () => '',
-      udcToken: () => {
-        return {
-          balance: Zero,
-        };
-      },
-    };
+    const confirmButton = wrapper
+      .findComponent(RaidenDialog)
+      .findAll('button')
+      .at(1);
+    confirmButton.trigger('click');
+    await wrapper.vm.$nextTick();
 
-    const mockStore = new Vuex.Store({ state, getters });
-    return mount(Withdrawal, {
-      vuetify,
-      store: mockStore,
-      stubs: ['v-dialog', 'i18n'],
-      mocks: {
-        $identicon: $identicon(),
-        $t: (msg: string) => msg,
-        $raiden,
-      },
-    });
-  }
-
-  beforeEach(() => {
-    vuetify = new Vuetify();
-    $raiden = new RaidenService(store) as Mocked<RaidenService>;
+    expect($raiden.transferOnChainTokens).toHaveBeenCalledTimes(1);
   });
 
-  describe('no tokens', () => {
-    beforeEach(() => {
-      $raiden.getRaidenAccountBalances.mockResolvedValue([]);
-      wrapper = createWrapper();
-    });
+  test('displays empty screen if there are no tokens to withdraw', async () => {
+    const wrapper = createWrapper(tokenNoBalance, raidenBalance);
+    await flushPromises();
 
-    test('show an empty screen', () => {
-      expect(wrapper.find('.withdrawal__empty').exists()).toBe(true);
-    });
+    expect(wrapper.find('.withdrawal__empty').exists()).toBe(true);
   });
 
-  describe('with tokens and no balance', () => {
-    beforeEach(() => {
-      $raiden.getRaidenAccountBalances.mockResolvedValue([
-        {
-          address: '0xtoken',
-          decimals: 5,
-          balance: parseUnits('1.2', 5),
-          name: 'TestToken',
-          symbol: 'TTT',
-        },
-      ]);
-      wrapper = createWrapper();
-    });
+  test('disaples withdrawal confirm button if Raiden account has no balance', async () => {
+    const wrapper = createWrapper(tokenBalance, raidenNoBalance);
+    await flushPromises();
 
-    test('open a dialog when clicking withdraw for a token', async () => {
-      wrapper.find('.withdrawal__tokens').find('button').trigger('click');
+    wrapper.find('.withdrawal__tokens__button').trigger('click');
+    await wrapper.vm.$nextTick();
 
-      await wrapper.vm.$nextTick();
+    const confirmButton = wrapper
+      .findComponent(RaidenDialog)
+      .findAll('button')
+      .at(1);
 
-      expect(wrapper.findComponent(RaidenDialog).exists()).toBe(true);
-    });
-
-    test('disables withdrawal if the Raiden account has no funds', async () => {
-      wrapper.find('.withdrawal__tokens').find('button').trigger('click');
-
-      await wrapper.vm.$nextTick();
-
-      expect(
-        wrapper.find('.withdrawal-dialog__action').find('button').attributes()[
-          'disabled'
-        ]
-      ).toBe('disabled');
-    });
-
-    test('withdraw is no-op without a selection', async () => {
-      await (wrapper.vm as any).withdrawTokens();
-      expect($raiden.transferOnChainTokens).toHaveBeenCalledTimes(0);
-    });
-  });
-
-  describe('with tokens and balance', () => {
-    beforeEach(() => {
-      $raiden.getRaidenAccountBalances.mockResolvedValue([
-        {
-          address: '0xtoken',
-          decimals: 5,
-          balance: parseUnits('1.2', 5),
-          name: 'TestToken',
-          symbol: 'TTT',
-        },
-      ]);
-      wrapper = createWrapper('2');
-    });
-
-    test('withdraws tokens when the user clicks on the action button', async () => {
-      wrapper.find('.withdrawal__tokens').find('button').trigger('click');
-
-      await wrapper.vm.$nextTick();
-
-      expect(
-        wrapper.find('.withdrawal-dialog__action').find('button').attributes()[
-          'disabled'
-        ]
-      ).toBeUndefined();
-
-      wrapper
-        .find('.withdrawal-dialog__action')
-        .find('button')
-        .trigger('click');
-
-      await wrapper.vm.$nextTick();
-      expect($raiden.transferOnChainTokens).toHaveBeenCalledTimes(1);
-      await wrapper.vm.$nextTick();
-      expect(wrapper.find('.withdrawal__empty').exists()).toBe(true);
-    });
+    expect(confirmButton.attributes()['disabled']).toBe('disabled');
   });
 });